+++ src/catalog_generated.c
@@ -818,7 +818,7 @@ static const roadcast_signal_part_t SIGNAL_PARTS_812[] = {{4, 24, 8}};
 static const roadcast_signal_part_t SIGNAL_PARTS_813[] = {{0, 24, 8}, {0, 16, 8}, {0, 8, 8}, {0, 0, 8}, {4, 24, 8}, {4, 16, 8}, {4, 8, 8}, {4, 0, 8}};
 static const roadcast_signal_part_t SIGNAL_PARTS_814[] = {{0, 24, 8}, {0, 16, 8}, {0, 8, 8}, {0, 0, 8}, {4, 24, 8}, {4, 16, 8}, {4, 8, 8}, {4, 0, 8}};
 
-const uint64_t ROADCAST_CAN_SCHEMA_HASH = UINT64_C(0xf4cd0f01fb221d14);
+const uint64_t ROADCAST_CAN_SCHEMA_HASH = UINT64_C(0xf8db8ab0f49a0a7f);
 
 const roadcast_signal_definition_t
     ROADCAST_SIGNALS[ROADCAST_SIGNAL_COUNT] = {
@@ -7245,7 +7245,7 @@ const roadcast_signal_definition_t
         {
             .stable_id = UINT64_C(0xfac0a4257f62ffc9),
             .name = "OBC_iInAct",
-            .unit = "",
+            .unit = "A",
             .parts = SIGNAL_PARTS_428,
             .invalid_signal_index = UINT32_C(4294967295),
             .can_id = UINT16_C(0x221),
@@ -7253,14 +7253,14 @@ const roadcast_signal_definition_t
             .part_count = UINT8_C(2),
             .width = UINT8_C(10),
             .is_signed = UINT8_C(0),
-            .calibrated = UINT8_C(0),
-            .scale = 1,
+            .calibrated = UINT8_C(1),
+            .scale = 0.10000000000000001,
             .offset = 0,
         },
         {
             .stable_id = UINT64_C(0x832e364e51198ffd),
             .name = "OBC_uInAct",
-            .unit = "",
+            .unit = "V",
             .parts = SIGNAL_PARTS_429,
             .invalid_signal_index = UINT32_C(4294967295),
             .can_id = UINT16_C(0x221),
@@ -7268,8 +7268,8 @@ const roadcast_signal_definition_t
             .part_count = UINT8_C(2),
             .width = UINT8_C(13),
             .is_signed = UINT8_C(0),
-            .calibrated = UINT8_C(0),
-            .scale = 1,
+            .calibrated = UINT8_C(1),
+            .scale = 0.10000000000000001,
             .offset = 0,
         },
         {